# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BarberApp API
 * OpenAPI spec version: 0.1.0
 */
export interface RescheduleByTokenInput {
  /** ISO 8601 datetime of the new slot */
  scheduledAt: string;
  /** Optionally switch barber when rescheduling */
  barberId?: number;
}

export interface HealthStatus {
  status: string;
}

export interface Client {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface ClientInput {
  name: string;
  phone: string;
  email?: string;
  notes?: string;
}

export interface ClientUpdate {
  name?: string;
  phone?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface Service {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  durationMinutes: number;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder: number;
  /** IDs of barbers that perform this service. Empty array means "all barbers" (legacy). */
  barberIds: number[];
}

export interface ServiceInput {
  name: string;
  description?: string;
  durationMinutes: number;
  price: number;
  imageUrl?: string;
  sortOrder?: number;
  barberIds?: number[];
}

export interface ServiceUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  durationMinutes?: number;
  price?: number;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder?: number;
  barberIds?: number[];
}

export interface DaySchedule {
  closed: boolean;
  open: string;
  close: string;
  lunchStart: string;
  lunchEnd: string;
}

export interface WeeklySchedule {
  monday: DaySchedule;
  tuesday: DaySchedule;
  wednesday: DaySchedule;
  thursday: DaySchedule;
  friday: DaySchedule;
  saturday: DaySchedule;
  sunday: DaySchedule;
}

export interface Barber {
  id: number;
  name: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  active: boolean;
  sortOrder: number;
  createdAt: string;
  /** IDs of services this barber performs. Empty means "all services". */
  serviceIds: number[];
  /** Per-barber working hours. Null means "use shop schedule". */
  weeklySchedule?: WeeklySchedule | null;
}

export interface BarberInput {
  name: string;
  photoUrl?: string;
  bio?: string;
  active?: boolean;
  sortOrder?: number;
  serviceIds?: number[];
  weeklySchedule?: WeeklySchedule | null;
}

export interface BarberUpdate {
  name?: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  active?: boolean;
  sortOrder?: number;
  serviceIds?: number[];
  weeklySchedule?: WeeklySchedule | null;
}

export type AppointmentPaymentMethod = typeof AppointmentPaymentMethod[keyof typeof AppointmentPaymentMethod];


export const AppointmentPaymentMethod = {
  now: 'now',
  on_site: 'on_site',
} as const;

export interface Appointment {
  id: number;
  /** @nullable */
  clientId?: number | null;
  clientName: string;
  /** @nullable */
  serviceId?: number | null;
  serviceName: string;
  /** @nullable */
  barberId?: number | null;
  /** @nullable */
  barberName?: string | null;
  servicePrice: number;
  serviceDuration: number;
  scheduledAt: string;
  status: string;
  paymentMethod?: AppointmentPaymentMethod;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  cancelToken?: string | null;
  createdAt?: string;
}

export type AppointmentInputPaymentMethod = typeof AppointmentInputPaymentMethod[keyof typeof AppointmentInputPaymentMethod];


export const AppointmentInputPaymentMethod = {
  now: 'now',
  on_site: 'on_site',
} as const;

export interface AppointmentInput {
  /** Shop owner user ID (required for public booking page) */
  shopId?: string;
  clientId?: number;
  clientName: string;
  serviceId?: number;
  serviceName: string;
  barberId?: number;
  barberName?: string;
  servicePrice: number;
  serviceDuration: number;
  scheduledAt: string;
  paymentMethod?: AppointmentInputPaymentMethod;
  notes?: string;
}

export interface AppointmentUpdate {
  clientName?: string;
  serviceName?: string;
  servicePrice?: number;
  serviceDuration?: number;
  scheduledAt?: string;
  status?: string;
  /** @nullable */
  notes?: string | null;
}

export type AvailabilitySlotsItem = {
  /** HH:MM */
  time: string;
  available: boolean;
};

export interface Availability {
  date: string;
  dayClosed: boolean;
  slots: AvailabilitySlotsItem[];
}

export interface QueueEntry {
  id: number;
  /** @nullable */
  appointmentId?: number | null;
  clientName: string;
  serviceName: string;
  servicePrice: number;
  serviceDuration: number;
  position: number;
  status: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  scheduledAt?: string | null;
  createdAt: string;
}

export interface QueueInput {
  clientName: string;
  serviceName: string;
  servicePrice: number;
  serviceDuration: number;
  notes?: string;
}

export interface DashboardSummary {
  appointmentsToday: number;
  appointmentsCompleted: number;
  appointmentsPending: number;
  monthlyRevenue: number;
  totalClients: number;
  queueCount: number;
  currentAppointment: Appointment | null;
  nextAppointment: Appointment | null;
}

export interface RevenueByService {
  serviceName: string;
  revenue: number;
  count: number;
}

export interface RevenueByDay {
  date: string;
  revenue: number;
  count: number;
}

export interface FinancialSummary {
  totalRevenue: number;
  totalAppointments: number;
  averageTicket: number;
  revenueByService: RevenueByService[];
  revenueByDay: RevenueByDay[];
}

export interface Settings {
  id: number;
  barbershopName: string;
  ownerName: string;
  /**
     * @maxLength 3000000
     * @nullable
     */
  logoUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  openTime?: string | null;
  /** @nullable */
  closeTime?: string | null;
  weeklySchedule?: WeeklySchedule | null;
  /** @nullable */
  bookingPageMessage?: string | null;
  paymentEnableNow?: boolean;
  paymentEnableOnSite?: boolean;
}

export interface SettingsUpdate {
  barbershopName?: string;
  ownerName?: string;
  /**
     * @maxLength 3000000
     * @nullable
     */
  logoUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  openTime?: string | null;
  /** @nullable */
  closeTime?: string | null;
  weeklySchedule?: WeeklySchedule | null;
  /** @nullable */
  bookingPageMessage?: string | null;
  paymentEnableNow?: boolean;
  paymentEnableOnSite?: boolean;
}

export type ListClientsParams = {
search?: string;
};

export type ListServicesParams = {
/**
 * Shop owner user ID (required for public booking page)
 */
shopId?: string;
};

export type ListAppointmentsParams = {
date?: string;
status?: string;
};

export type GetAvailabilityParams = {
/**
 * Shop owner user ID (required for public booking page)
 */
shopId?: string;
/**
 * YYYY-MM-DD
 */
date: string;
serviceId: number;
/**
 * Filter availability to the agenda of one barber. If omitted, conflicts are checked across all appointments.
 */
barberId?: number;
};

export type GetFinancialSummaryParams = {
month?: number;
year?: number;
/**
 * Optional day of month (1-31). When provided, the summary is limited to that single day.
 */
day?: number;
};

export type ListBarbersParams = {
/**
 * Shop owner user ID (required for public booking page)
 */
shopId?: string;
activeOnly?: boolean;
/**
 * Only return barbers that perform this service
 */
serviceId?: number;
};

export type GetSettingsParams = {
/**
 * Shop owner user ID (required for public booking page)
 */
shopId?: string;
};