# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BarberApp API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /barbers:
    get:
      operationId: listBarbers
      tags: [barbers]
      summary: List barbers
      parameters:
        - name: shopId
          in: query
          description: Shop owner user ID (required for public booking page)
          schema:
            type: string
        - name: activeOnly
          in: query
          schema:
            type: boolean
        - name: serviceId
          in: query
          description: Only return barbers that perform this service
          schema:
            type: integer
      responses:
        "200":
          description: List of barbers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Barber"
    post:
      operationId: createBarber
      tags: [barbers]
      summary: Create a barber
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BarberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Barber"
  /barbers/{id}:
    get:
      operationId: getBarber
      tags: [barbers]
      summary: Get a barber
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Barber
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Barber"
    patch:
      operationId: updateBarber
      tags: [barbers]
      summary: Update a barber
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BarberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Barber"
    delete:
      operationId: deleteBarber
      tags: [barbers]
      summary: Delete a barber
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Client:
      type: object
      required: [id, name, phone, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ClientInput:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        notes:
          type: string
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    Barber:
      type: object
      required: [id, name, active, sortOrder, createdAt, serviceIds]
      properties:
        id:
          type: integer
        name:
          type: string
        photoUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        active:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
        serviceIds:
          type: array
          description: IDs of services this barber performs. Empty means "all services".
          items:
            type: integer
        weeklySchedule:
          description: Per-barber working hours. Null means "use shop schedule".
          oneOf:
            - $ref: "#/components/schemas/WeeklySchedule"
            - type: "null"
    BarberInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        photoUrl:
          type: string
        bio:
          type: string
        active:
          type: boolean
        sortOrder:
          type: integer
        serviceIds:
          type: array
          items:
            type: integer
        weeklySchedule:
          oneOf:
            - $ref: "#/components/schemas/WeeklySchedule"
            - type: "null"
    BarberUpdate:
      type: object
      properties:
        name:
          type: string
        photoUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        active:
          type: boolean
        sortOrder:
          type: integer
        serviceIds:
          type: array
          items:
            type: integer
        weeklySchedule:
          oneOf:
            - $ref: "#/components/schemas/WeeklySchedule"
            - type: "null"
...
    DaySchedule:
      type: object
      required: [closed, open, close, lunchStart, lunchEnd]
      properties:
        closed:
          type: boolean
        open:
          type: string
        close:
          type: string
        lunchStart:
          type: string
        lunchEnd:
          type: string
    WeeklySchedule:
      type: object
      required: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
      properties:
        monday:
          $ref: "#/components/schemas/DaySchedule"
        tuesday:
          $ref: "#/components/schemas/DaySchedule"
        wednesday:
          $ref: "#/components/schemas/DaySchedule"
        thursday:
          $ref: "#/components/schemas/DaySchedule"
        friday:
          $ref: "#/components/schemas/DaySchedule"
        saturday:
          $ref: "#/components/schemas/DaySchedule"
        sunday:
          $ref: "#/components/schemas/DaySchedule"
...